rename classe to class
