+++ src/client/app/components/class/class.ts
